Added new class for a helical curve, biosavart subclasses magneticfield
