+++ setup.py
@@ -16,10 +16,6 @@ PLAT_TO_CMAKE = {
     "win-arm64": "ARM64",
 }
 
-import site
-site.ENABLE_USER_SITE = True
-
-
 # A CMakeExtension needs a sourcedir instead of a file list.
 # The name must be the _single_ output extension from the CMake build.
 # If you need multiple extensions, see scikit-build.
